requirements.txt
flask
pandas
numpy
scikit-learn
joblib
openai
python-dotenv
